+++ components/music/musicTierVote.tsx
@@ -5,96 +5,18 @@ import {
     submitChartEvaluation,
     toggleChartEvaluationReaction,
 } from "@/app/(nevigation)/music/[index]/[difficulty]/action";
-import { cn } from "@/lib/utils";
-import { Check, Minus, Plus, ThumbsDown, ThumbsUp, Trash2 } from "lucide-react";
-import Link from "next/link";
 import { useRouter } from "next/navigation";
-import { Controller, useForm, useWatch } from "react-hook-form";
-import { useMemo, useState, useTransition } from "react";
-import {
-    Line,
-    LineChart,
-    ResponsiveContainer,
-    Tooltip,
-    XAxis,
-    YAxis,
-} from "recharts";
+import { useForm, useWatch } from "react-hook-form";
+import { useState, useTransition } from "react";
+import MusicEvaluationForm from "./musicEvaluationForm";
+import MusicOpinionList from "./musicOpinionList";
+import MusicTierSummary from "./musicTierSummary";
+import type {
+    EvaluationFormValues,
+    MusicTierVoteProps,
+} from "./musicTierVoteTypes";
 
-const patternItems = [
-    { key: "stairs", label: "계단" },
-    { key: "chord", label: "동치" },
-    { key: "trill", label: "트릴" },
-    { key: "glissando", label: "글리산도" },
-    { key: "repetition", label: "연타" },
-] as const;
-const patternLevels = ["없음", "낮음", "보통", "높음", "매우높음"];
-
-interface EvaluationFormValues {
-    perceivedConstant: number;
-    stairs: number | null;
-    chord: number | null;
-    trill: number | null;
-    glissando: number | null;
-    repetition: number | null;
-    comment: string;
-}
-
-interface MusicTierVoteProps {
-    chartId: number;
-    canVote: boolean;
-    difficulty: "Normal" | "Hard" | "Expert" | "Real";
-    level: number;
-    officialConstant: number | null;
-    tierConstant: number | null;
-    constantHistory: {
-        id: number;
-        value: number;
-        effectiveAt: string;
-    }[];
-    community: {
-        average: number | null;
-        count: number;
-        distribution: { value: number; count: number }[];
-    };
-    currentEvaluation: {
-        perceived_constant: number;
-        stairs: number;
-        chord: number;
-        trill: number;
-        glissando: number;
-        repetition: number;
-        comment: string | null;
-    } | null;
-    opinionCount: number;
-    opinions: {
-        id: number;
-        perceivedConstant: number;
-        comment: string;
-        updatedAt: string;
-        user: { id: number; username: string | null };
-        positiveCount: number;
-        negativeCount: number;
-        viewerReaction: number | null;
-        canDelete: boolean;
-    }[];
-}
-
-function formatMonth(value: string) {
-    return value.slice(0, 7).replace("-", ".");
-}
-
-function formatDate(value: string) {
-    return new Intl.DateTimeFormat("ko-KR", {
-        month: "2-digit",
-        day: "2-digit",
-    })
-        .format(new Date(value))
-        .replaceAll(". ", ".")
-        .replace(/\.$/, "")
-        .trim();
-}
-
-// 악곡별 서열 추이와 사용자 투표 UI를 한곳에서 관리함
+// 악곡별 서열 투표 상태와 서버 액션을 한곳에서 관리함
 export default function MusicTierVote({
     chartId,
     canVote,
@@ -143,48 +65,6 @@ export default function MusicTierVote({
         control,
         name: "perceivedConstant",
     });
-
-    const histogram = useMemo(() => {
-        if (community.distribution.length === 0) return [];
-
-        const distribution = new Map(
-            community.distribution.map((item) => [
-                Math.round(item.value * 10),
-                item.count,
-            ])
-        );
-        const values = [...distribution.keys()];
-        const minimum = Math.min(...values);
-        const maximum = Math.max(...values);
-
-        return Array.from({ length: maximum - minimum + 1 }, (_, index) => {
-            const tick = minimum + index;
-            const value = tick / 10;
-            const count = distribution.get(tick) ?? 0;
-
-            return { value, count };
-        });
-    }, [community.distribution]);
-    const maximumHistogramCount = Math.max(
-        1,
-        ...histogram.map((item) => item.count)
-    );
-    const highlightedHistogramValue =
-        community.average === null || community.distribution.length === 0
-            ? null
-            : community.distribution.reduce((closest, item) =>
-                  Math.abs(item.value - community.average!) <
-                  Math.abs(closest.value - community.average!)
-                      ? item
-                      : closest
-              ).value;
-    const constantDelta =
-        constantHistory.length > 1
-            ? constantHistory.at(-1)!.value - constantHistory[0].value
-            : 0;
-    const hasPatternError = patternItems.some(
-        ({ key }) => errors[key] !== undefined
-    );
     const parsedPerceivedConstant = Number(perceivedInput);
     const isPerceivedConstantValid =
         Number.isFinite(parsedPerceivedConstant) &&
@@ -293,457 +173,34 @@ export default function MusicTierVote({
 
     return (
         <div className="flex flex-col gap-3">
-            <section className="bg-surface rounded-card p-4">
-                <header className="flex items-center gap-2">
-                    <h2 className="text-section">서열 상수 변동</h2>
-                    {tierConstant !== null ? (
-                        <strong className="text-text-primary text-base tabular-nums">
-                            {tierConstant.toFixed(1)}
-                        </strong>
-                    ) : null}
-                    {constantDelta !== 0 ? (
-                        <span
-                            className={cn(
-                                "rounded px-1.5 py-1 text-xs font-bold tabular-nums",
-                                constantDelta > 0
-                                    ? "bg-danger/15 text-danger"
-                                    : "bg-chart/15 text-chart"
-                            )}
-                        >
-                            {constantDelta > 0 ? "+" : ""}
-                            {constantDelta.toFixed(1)}
-                        </span>
-                    ) : null}
-                </header>
+            <MusicTierSummary
+                tierConstant={tierConstant}
+                constantHistory={constantHistory}
+                community={community}
+            />
 
-                {constantHistory.length > 0 ? (
-                    <div className="mt-2 h-24">
-                        <ResponsiveContainer width="100%" height="100%">
-                            <LineChart
-                                data={constantHistory}
-                                margin={{
-                                    top: 12,
-                                    right: 8,
-                                    bottom: 14,
-                                    left: 8,
-                                }}
-                            >
-                                <XAxis dataKey="effectiveAt" hide />
-                                <YAxis
-                                    domain={["dataMin - 0.1", "dataMax + 0.1"]}
-                                    hide
-                                />
-                                <Tooltip
-                                    contentStyle={{
-                                        background:
-                                            "var(--color-surface-muted)",
-                                        border: "1px solid var(--color-border)",
-                                        borderRadius: 6,
-                                        fontSize: 12,
-                                    }}
-                                    formatter={(value) => [
-                                        Number(value).toFixed(1),
-                                        "서열 상수",
-                                    ]}
-                                    labelFormatter={(value) =>
-                                        formatMonth(String(value))
-                                    }
-                                />
-                                <Line
-                                    type="linear"
-                                    dataKey="value"
-                                    stroke="var(--color-text-secondary)"
-                                    strokeWidth={3}
-                                    dot={{
-                                        r: 3,
-                                        fill: "var(--color-text-disabled)",
-                                        strokeWidth: 0,
-                                    }}
-                                    activeDot={{
-                                        r: 5,
-                                        fill: "var(--color-text-primary)",
-                                        stroke: "var(--color-border)",
-                                        strokeWidth: 3,
-                                    }}
-                                />
-                            </LineChart>
-                        </ResponsiveContainer>
-                        <div className="text-text-disabled -mt-4 flex justify-between text-xs">
-                            <span>
-                                {formatMonth(constantHistory[0].effectiveAt)}{" "}
-                                등재
-                            </span>
-                            <span>
-                                {formatMonth(
-                                    constantHistory.at(-1)!.effectiveAt
-                                )}{" "}
-                                현재
-                            </span>
-                        </div>
-                    </div>
-                ) : (
-                    <p className="text-body-muted flex h-20 items-center justify-center">
-                        등록된 서열 상수 이력이 없습니다.
-                    </p>
-                )}
-            </section>
+            <MusicEvaluationForm
+                canVote={canVote}
+                isPending={isPending}
+                hasCurrentEvaluation={Boolean(currentEvaluation)}
+                isCommentExpanded={isCommentExpanded}
+                message={message}
+                control={control}
+                errors={errors}
+                perceivedConstantField={perceivedConstantField}
+                commentField={commentField}
+                onChangeConstant={changeConstant}
+                onCommentExpandedChange={setIsCommentExpanded}
+                onSubmit={handleSubmit(submitVote)}
+            />
 
-            <section className="bg-surface rounded-card p-4">
-                <header className="flex items-baseline gap-2">
-                    <h2 className="text-section">커뮤니티 체감</h2>
-                    <strong className="text-real text-xl font-extrabold tabular-nums">
-                        {community.average?.toFixed(2) ?? "-"}
-                    </strong>
-                    <span className="text-caption ml-auto">
-                        투표 {community.count.toLocaleString("ko-KR")}
-                    </span>
-                </header>
-                {histogram.length > 0 ? (
-                    <div className="mt-3 overflow-x-auto pb-1">
-                        <div
-                            className={cn(
-                                "flex h-14 items-end gap-1",
-                                histogram.length <= 6
-                                    ? "justify-center"
-                                    : "w-max"
-                            )}
-                        >
-                            {histogram.map((item) => (
-                                <div
-                                    key={item.value}
-                                    className="flex h-full w-12 shrink-0 flex-col items-center justify-end gap-1"
-                                >
-                                    <span
-                                        className={cn(
-                                            "min-h-1 w-full rounded-t-sm",
-                                            highlightedHistogramValue !==
-                                                null &&
-                                                Math.round(item.value * 10) ===
-                                                    Math.round(
-                                                        highlightedHistogramValue *
-                                                            10
-                                                    )
-                                                ? "bg-real"
-                                                : "bg-border"
-                                        )}
-                                        style={{
-                                            height: `${Math.max(
-                                                item.count > 0 ? 10 : 4,
-                                                (item.count /
-                                                    maximumHistogramCount) *
-                                                    34
-                                            )}px`,
-                                        }}
-                                    />
-                                    <span className="text-text-disabled text-xs tabular-nums">
-                                        {item.value.toFixed(1)}
-                                    </span>
-                                </div>
-                            ))}
-                        </div>
-                    </div>
-                ) : (
-                    <p className="text-body-muted flex h-14 items-center justify-center">
-                        아직 등록된 체감 난이도 투표가 없습니다.
-                    </p>
-                )}
-            </section>
-
-            <section className="bg-surface rounded-card p-4">
-                <h2 className="text-section">체감 난이도 투표</h2>
-                {!canVote ? (
-                    <p className="bg-danger/10 text-danger mt-3 rounded-md px-3 py-2 text-xs font-medium">
-                        해당 채보의 플레이 기록 연동 후 투표할 수 있습니다.
-                    </p>
-                ) : null}
-                <div
-                    className={cn(
-                        "mt-3 gap-2",
-                        isCommentExpanded ? "block" : "flex"
-                    )}
-                >
-                    {!isCommentExpanded ? (
-                        <div
-                            className={cn(
-                                "border-border flex h-10 shrink-0 overflow-hidden rounded-lg border",
-                                errors.perceivedConstant &&
-                                    "border-danger ring-danger ring-1"
-                            )}
-                        >
-                            <button
-                                type="button"
-                                disabled={!canVote}
-                                aria-label="체감 난이도 0.1 낮추기"
-                                className="text-text-secondary hover:bg-surface-muted flex w-10 items-center justify-center disabled:opacity-40"
-                                onClick={() => changeConstant(-0.1)}
-                            >
-                                <Minus size={16} aria-hidden />
-                            </button>
-                            <input
-                                {...perceivedConstantField}
-                                type="number"
-                                min="1"
-                                max="14"
-                                step="0.1"
-                                inputMode="decimal"
-                                disabled={!canVote}
-                                aria-label="체감 난이도 직접 입력"
-                                aria-invalid={Boolean(errors.perceivedConstant)}
-                                className="bg-surface-muted text-text-primary h-full w-16 appearance-none px-1 text-center text-base font-extrabold tabular-nums outline-none disabled:opacity-40 [&::-webkit-inner-spin-button]:appearance-none [&::-webkit-outer-spin-button]:appearance-none"
-                            />
-                            <button
-                                type="button"
-                                disabled={!canVote}
-                                aria-label="체감 난이도 0.1 높이기"
-                                className="text-text-secondary hover:bg-surface-muted flex w-10 items-center justify-center disabled:opacity-40"
-                                onClick={() => changeConstant(0.1)}
-                            >
-                                <Plus size={16} aria-hidden />
-                            </button>
-                        </div>
-                    ) : null}
-                    <textarea
-                        {...commentField}
-                        disabled={!canVote}
-                        maxLength={120}
-                        rows={isCommentExpanded ? 4 : 1}
-                        placeholder="짧은 코멘트"
-                        aria-label="체감 난이도 코멘트"
-                        aria-invalid={Boolean(errors.comment)}
-                        onFocus={() => setIsCommentExpanded(true)}
-                        onBlur={(event) => {
-                            commentField.onBlur(event);
-                            setIsCommentExpanded(false);
-                        }}
-                        className={cn(
-                            "border-border bg-bg text-input placeholder:text-text-disabled focus:border-text-secondary min-w-0 flex-1 resize-none rounded-lg border border-dashed px-3 py-2 transition-[height] outline-none disabled:opacity-40",
-                            isCommentExpanded
-                                ? "h-24 w-full overflow-y-auto"
-                                : "h-10 overflow-hidden",
-                            errors.comment && "border-danger ring-danger ring-1"
-                        )}
-                    />
-                    {!isCommentExpanded ? (
-                        <button
-                            type="button"
-                            disabled={isPending || !canVote}
-                            onClick={handleSubmit(submitVote)}
-                            className="bg-text-primary text-bg flex h-10 shrink-0 items-center gap-1 rounded-lg px-3 text-sm font-bold disabled:opacity-50"
-                        >
-                            {currentEvaluation ? (
-                                <Check size={14} aria-hidden />
-                            ) : null}
-                            {isPending
-                                ? "처리 중"
-                                : currentEvaluation
-                                  ? "수정"
-                                  : "제출"}
-                        </button>
-                    ) : null}
-                </div>
-                {errors.perceivedConstant ? (
-                    <p className="text-danger mt-1 text-xs" role="alert">
-                        {errors.perceivedConstant.message}
-                    </p>
-                ) : null}
-                {errors.comment ? (
-                    <p className="text-danger mt-1 text-xs" role="alert">
-                        {errors.comment.message}
-                    </p>
-                ) : null}
-
-                <div className="border-divider mt-4 border-t pt-3">
-                    <div className="flex items-baseline gap-2">
-                        <h3 className="text-text-primary text-sm font-bold">
-                            패턴 투표
-                        </h3>
-                        <span className="text-caption">
-                            필수 · 제출 시 함께 반영
-                        </span>
-                    </div>
-                    <div className="mt-3 grid grid-cols-[4rem_repeat(5,minmax(0,1fr))] items-center">
-                        <span />
-                        {patternLevels.map((label) => (
-                            <span
-                                key={label}
-                                className="text-text-disabled text-center text-xs"
-                            >
-                                {label}
-                            </span>
-                        ))}
-                        {patternItems.map(({ key, label }) => (
-                            <Controller
-                                key={key}
-                                name={key}
-                                control={control}
-                                rules={{
-                                    validate: (value) =>
-                                        value !== null ||
-                                        "다섯 패턴 항목을 모두 선택해 주세요.",
-                                }}
-                                render={({ field }) => (
-                                    <div className="contents">
-                                        <span className="text-text-secondary text-sm">
-                                            {label}
-                                        </span>
-                                        {patternLevels.map((option, value) => (
-                                            <label
-                                                key={option}
-                                                className={cn(
-                                                    "flex h-9 items-center justify-center",
-                                                    canVote
-                                                        ? "cursor-pointer"
-                                                        : "cursor-not-allowed opacity-40"
-                                                )}
-                                                title={`${label} ${option}`}
-                                            >
-                                                <input
-                                                    ref={
-                                                        value === 0
-                                                            ? field.ref
-                                                            : undefined
-                                                    }
-                                                    type="radio"
-                                                    name={field.name}
-                                                    value={value}
-                                                    checked={
-                                                        field.value === value
-                                                    }
-                                                    disabled={!canVote}
-                                                    onBlur={field.onBlur}
-                                                    onChange={() =>
-                                                        field.onChange(value)
-                                                    }
-                                                    className="sr-only"
-                                                />
-                                                <span
-                                                    className={cn(
-                                                        "size-4 rounded-full border transition-colors",
-                                                        field.value === value
-                                                            ? "border-chart bg-chart"
-                                                            : "border-border",
-                                                        errors[key] &&
-                                                            "border-danger"
-                                                    )}
-                                                />
-                                            </label>
-                                        ))}
-                                    </div>
-                                )}
-                            />
-                        ))}
-                    </div>
-                    {hasPatternError ? (
-                        <p className="text-danger mt-2 text-xs" role="alert">
-                            다섯 패턴 항목을 모두 선택해 주세요.
-                        </p>
-                    ) : null}
-                </div>
-                {message ? (
-                    <p
-                        className="text-caption mt-3"
-                        role="status"
-                        aria-live="polite"
-                    >
-                        {message}
-                    </p>
-                ) : null}
-            </section>
-
-            <section className="bg-surface rounded-card overflow-hidden">
-                <header className="bg-surface-muted flex h-10 items-center px-4">
-                    <h2 className="text-section">
-                        의견 {opinionCount.toLocaleString("ko-KR")}
-                    </h2>
-                </header>
-                {opinions.length > 0 ? (
-                    <ol>
-                        {opinions.map((opinion) => (
-                            <li
-                                key={opinion.id}
-                                className="border-divider border-t px-4 py-3 first:border-t-0"
-                            >
-                                <div className="flex items-center gap-2">
-                                    <Link
-                                        href={`/profile/${opinion.user.id}`}
-                                        className="text-text-primary text-sm font-bold"
-                                    >
-                                        {opinion.user.username ||
-                                            "이름 없는 유저"}
-                                    </Link>
-                                    <span className="text-caption tabular-nums">
-                                        체감{" "}
-                                        {opinion.perceivedConstant.toFixed(1)}
-                                    </span>
-                                    <time
-                                        className="text-caption"
-                                        dateTime={opinion.updatedAt}
-                                    >
-                                        {formatDate(opinion.updatedAt)}
-                                    </time>
-                                </div>
-                                <p className="text-body mt-1">
-                                    {opinion.comment}
-                                </p>
-                                <div className="mt-2 flex items-center gap-1">
-                                    <button
-                                        type="button"
-                                        disabled={isPending}
-                                        onClick={() =>
-                                            reactToOpinion(opinion.id, 1)
-                                        }
-                                        aria-label="의견 추천"
-                                        className={cn(
-                                            "flex h-8 items-center gap-1 rounded-md px-2 text-xs",
-                                            opinion.viewerReaction === 1
-                                                ? "bg-chart/15 text-chart"
-                                                : "text-text-secondary hover:bg-surface-muted"
-                                        )}
-                                    >
-                                        <ThumbsUp size={14} aria-hidden />
-                                        {opinion.positiveCount}
-                                    </button>
-                                    <button
-                                        type="button"
-                                        disabled={isPending}
-                                        onClick={() =>
-                                            reactToOpinion(opinion.id, -1)
-                                        }
-                                        aria-label="의견 비추천"
-                                        className={cn(
-                                            "flex h-8 items-center gap-1 rounded-md px-2 text-xs",
-                                            opinion.viewerReaction === -1
-                                                ? "bg-danger/15 text-danger"
-                                                : "text-text-secondary hover:bg-surface-muted"
-                                        )}
-                                    >
-                                        <ThumbsDown size={14} aria-hidden />
-                                        {opinion.negativeCount}
-                                    </button>
-                                    {opinion.canDelete ? (
-                                        <button
-                                            type="button"
-                                            disabled={isPending}
-                                            onClick={() =>
-                                                deleteEvaluation(opinion.id)
-                                            }
-                                            aria-label="내 투표와 의견 삭제"
-                                            title="내 투표와 의견 삭제"
-                                            className="text-text-secondary hover:bg-danger/15 hover:text-danger ml-auto flex size-8 items-center justify-center rounded-md disabled:opacity-50"
-                                        >
-                                            <Trash2 size={14} aria-hidden />
-                                        </button>
-                                    ) : null}
-                                </div>
-                            </li>
-                        ))}
-                    </ol>
-                ) : (
-                    <p className="text-body-muted flex h-24 items-center justify-center">
-                        아직 등록된 의견이 없습니다.
-                    </p>
-                )}
-            </section>
+            <MusicOpinionList
+                opinionCount={opinionCount}
+                opinions={opinions}
+                isPending={isPending}
+                onReact={reactToOpinion}
+                onDelete={deleteEvaluation}
+            />
         </div>
     );
 }
