+++ components/music/musicTierVoteConfig.ts
@@ -1,0 +1,9 @@
+export const patternItems = [
+    { key: "stairs", label: "계단" },
+    { key: "chord", label: "동치" },
+    { key: "trill", label: "트릴" },
+    { key: "glissando", label: "글리산도" },
+    { key: "repetition", label: "연타" },
+] as const;
+
+export const patternLevels = ["없음", "낮음", "보통", "높음", "매우높음"];
